skip odd numbers from the switch
After some research, I found that no odd number can be equal to 104,
so we can exit early if the number is odd and skip the switch statement.
I preferred to use the external library 'is-odd' to check if a number
is odd instead of open coding it. This is a good practice in JS and
improves code readability.

Tests demonstrates not only that the check now is blazing fast for
odd numbers, but also small even numbers give a ~50% speed increase.